# server/docs/openapi/spec.yaml
...
openapi: 3.1.1
info:
  title: Forms Portal API
  version: 0.1.0
  summary: API Documentation for the Forms Portal
  contact:
    name: OSDG
servers:
  - url: /api
    description: Production
...
security:
  - cookieAuth: []
paths:
...
  /forms/{formId}/responses:
    parameters:
      - $ref: '#/components/parameters/formId'
    get:
      tags: [Responses]
      summary: List form responses
      description: Retrieves all completed and in-progress responses for a form. Requires ANALYZE permission or ownership of responses.
      operationId: listResponses
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ResponseStatus'
            default: completed
        - name: sort
          in: query
          description: Sort criteria.
          schema:
            type: string
            enum: [started, submitted]
            default: submitted
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of form responses.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Response'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Responses]
      summary: Start response
      description: Initiates a new response session for the form. Requires RESPOND permission.
      operationId: startResponse
      responses:
        '201':
          description: Response session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /forms/{formId}/responses/{responseId}:
    parameters:
      - $ref: '#/components/parameters/formId'
      - $ref: '#/components/parameters/responseId'
    get:
      tags: [Responses]
      summary: Get response details
      description: Retrieves a specific response. Requires ANALYZE permission or ownership of the response.
      operationId: getResponse
      responses:
        '200':
          description: Response details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /forms/{formId}/responses/{responseId}/answers:
    parameters:
      - $ref: '#/components/parameters/formId'
      - $ref: '#/components/parameters/responseId'
    get:
      tags: [Responses]
      summary: Get answers
      description: Retrieves all the answers submitted as part of a specific response. Requires ANALYZE permission or ownership of the response.
      operationId: getAnswers
      responses:
        '200':
          description: Response details.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Answer'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    put:
      tags: [Responses]
      summary: Save answer
      description: Creates or updates an answer for a specific question within a response. Requires ownership of response.
      operationId: saveAnswer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerUpsert'
      responses:
        '200':
          description: Answer saved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Answer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /forms/{formId}/responses/{responseId}/submit:
    parameters:
      - $ref: '#/components/parameters/formId'
      - $ref: '#/components/parameters/responseId'
    post:
      tags: [Responses]
      summary: Submit response
      description: Finalizes a response by changing its status to 'completed'. Requires ownership of response.
      operationId: submitResponse
      responses:
        '200':
          description: Response submitted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    cookieAuth:
      type: cookie
      name: session
  parameters:
...
    formId:
      name: formId
      in: path
      required: true
      schema:
        type: string
        format: ulid
    responseId:
      name: responseId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    UnprocessableEntity:
      description: Unprocessable Entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
  schemas:
...
    ResponseStatus:
      type: string
      enum: [draft, completed, edited]
...
    Response:
      type: object
      required:
        - id
        - form
        - status
        - started
      properties:
        id:
          type: string
        form:
          type: string
          format: ulid
        respondent:
          type: string
          format: ulid
          nullable: true
        status:
          $ref: '#/components/schemas/ResponseStatus'
        started:
          type: string
          format: date-time
        submitted:
          type: string
          format: date-time
          nullable: true
        edited:
          type: string
          format: date-time
          nullable: true
    Answer:
      type: object
      required:
        - id
        - response
        - question
        - value
        - submitted
        - modified
      properties:
        id:
          type: string
          format: ulid
        response:
          type: string
        question:
          type: string
        value:
          type: object
          additionalProperties: true
        submitted:
          type: string
          format: date-time
        modified:
          type: string
          format: date-time
    AnswerUpsert:
      type: object
      required:
        - question
        - value
      properties:
        question:
          type: string
        value:
          type: object
          additionalProperties: true
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    ValidationError:
      type: object
      required:
        - code
        - message
        - errors
      properties:
        code:
          type: string
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...